Refactor exports + document barrels
