+++ packages/crystallis/lib/generated.dart
@@ -1,0 +1,14 @@
+/// Crystallis runtime for generated code.
+///
+/// This library is meant to be imported by generated `.data.g.dart` files,
+/// not by user code directly.
+library;
+
+export 'src/mixins/base.dart';
+export 'src/mixins/mutable.dart';
+export 'src/mixins/immutable.dart';
+export 'src/mixins/copyable.dart';
+export 'src/field_metadata.dart';
+export 'api/validator.dart';
+export 'src/utils.dart';
+export 'src/crystallise_annotation.dart';
