+++ packages/crystallis/lib/serialization.dart
@@ -1,0 +1,8 @@
+/// Crystallis serialization for generated code.
+///
+/// This library is meant to be imported by generated `.data.g.dart` files
+/// when serialization is enabled.
+library;
+
+export 'src/serializer.dart';
+export 'api/serializer.dart';
