+++ packages/crystallis/lib/src/utils.dart
@@ -1,0 +1,3 @@
+// External dependencies
+export 'package:meta/meta.dart' show immutable;
+export 'package:collection/collection.dart' show DeepCollectionEquality;
