+++ packages/crystallis/lib/src/validators/regex.dart
@@ -1,6 +1,6 @@
 import 'dart:core';
 
-import '../runtime/validator.dart';
+import '../../api/validator.dart';
 
 // ignore: public_member_api_docs
 class RegEx extends Validator {
